Updates to genutils, gitutils, and spreadsheet utils
MODIFIED FILES:
    jllutils/genutils.py - new functions, "chrange()" which creates a character
range, and "flatten()" which is intended to allow one to squash nested lists to
a flat (unnested) list, but still working on how to do that.

    jllutils/gitutils.py - filter_unlisted() now calls _make_relative_to_git_dir
once only for keep_files and ensures that all files in all branches are checked
to see if they should be kept. _make_relative_to_git_dir() now, by default, asks
what to do if it cannot make a path relative to the git directory; this helps if
the file does not exist in the current branch. This can be overridden, in which
case it will error if it cannot find the file. _iter_git_branches() now has a
"short" option that uses basename() to cut out the refs/heads part; this should
be redone to allow for slashes in branch names.

    jllutils/spreadsheet.py - save_range_as_csv() has "skip_empty" parameter
added; it's passed through tp iter_cell_range(). read_range() returns a cell
range as a list-of-lists. iter_cell_range() split out into its own function.